requirements.txt
streamlit>=1.31.0
openai>=1.12.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
numpy<2.0.0
tiktoken>=0.6.0
